[material-sample] Making material app from the basics
